Moved fmd management commands to sdk
